Add the functionality of using hexadecimal strings to specify the color (#44)
* add icon URL to the documentation

* add the functionality of using hexadecimal strings to specify the color

added functionality allows the user to either specify a color as an
integer or (much easier) allows the user to specify a hexadecimal value
in the form of the string as the color. If the user specifies both
color and hex_color as arguments then it should default to use the color
argument.